Update README.md
episodes had flaws. The episodes lacked the serial nature of the show because they lacked a sort of case-of-the-week feel. At first, the characters were interesting, but as the episodes continued, no attempt was made to delve deeper, so they grew progressively less interesting. The authors of the story were willing to have their main characters act foolishly in order to get what they wanted.